+++ src/main/java/com/moonspoon/moonspoon/dto/request/workbook/WorkbookCreateRequest.java
@@ -1,4 +1,4 @@
-package com.moonspoon.moonspoon.dto.request;
+package com.moonspoon.moonspoon.dto.request.workbook;
 
 import com.moonspoon.moonspoon.domain.Workbook;
 import jakarta.validation.constraints.NotBlank;
@@ -9,18 +9,20 @@ import java.time.LocalDateTime;
 
 @Getter
 @Setter
-public class WorkbookUpdateRequest {
+public class WorkbookCreateRequest {
 
     @NotBlank(message = "문제집 이름은 필수 입력입니다.")
     private String title;
     private String content;
-    private LocalDateTime updateDate;
+    private LocalDateTime createDate;
 
-    public static Workbook toEntity(WorkbookUpdateRequest dto){
+    public static Workbook toEntity(WorkbookCreateRequest dto){
         return Workbook.builder()
                 .title(dto.getTitle())
                 .content(dto.getContent())
-                .updateDate(dto.getUpdateDate())
+                .createDate(dto.getCreateDate())
                 .build();
     }
+
+
 }
