+++ src/main/java/com/moonspoon/moonspoon/service/ProblemService.java
@@ -3,8 +3,8 @@ package com.moonspoon.moonspoon.service;
 import com.moonspoon.moonspoon.domain.Problem;
 import com.moonspoon.moonspoon.domain.User;
 import com.moonspoon.moonspoon.domain.Workbook;
-import com.moonspoon.moonspoon.dto.request.ProblemCreateRequest;
-import com.moonspoon.moonspoon.dto.request.ProblemUpdateRequest;
+import com.moonspoon.moonspoon.dto.request.problem.ProblemCreateRequest;
+import com.moonspoon.moonspoon.dto.request.problem.ProblemUpdateRequest;
 import com.moonspoon.moonspoon.dto.response.ProblemResponse;
 import com.moonspoon.moonspoon.exception.NotFoundException;
 import com.moonspoon.moonspoon.exception.NotUserException;
