+++ src/main/java/com/moonspoon/moonspoon/service/UserService.java
@@ -1,6 +1,6 @@
 package com.moonspoon.moonspoon.service;
 import com.moonspoon.moonspoon.domain.User;
-import com.moonspoon.moonspoon.dto.request.UserSignupRequest;
+import com.moonspoon.moonspoon.dto.request.user.UserSignupRequest;
 import com.moonspoon.moonspoon.dto.response.UserResponse;
 import com.moonspoon.moonspoon.exception.DuplicateUserException;
 import com.moonspoon.moonspoon.repository.UserRepository;
